Move grant/default.go to output/grant.go
There's only one grant type now, so there's no need for a "default" grant
